python-iac/bin/requirements.txt
atlassian-python-api
aws_encryption_sdk
boto3
bson
cloudflare
cryptography
docker
GitPython
kubernetes==22.6.0
psycopg2-binary
pymongo
pyyaml
requests
sqlalchemy
sqlalchemy-redshift
python-dotenv





